[Update] 规范 mongodb 操作代码

+++ microservices/database/mongodb/pandas_opration.py
@@ -6,49 +6,49 @@ created by Fangyang on Time:2018/10/28
 __author__ = 'Fangyang'
 
 import pandas as pd
-from pymongo import MongoClient
 import json
 import time
+
+from typing import Tuple
 from tqdm import tqdm
+from pymongo import MongoClient
 
 
 class MongoDB(object):
 
-    def __init__(self, host: str = 'localhost', port: int = 27017, username=None, password=None):
+    def __init__(self, host: str = 'localhost', port: int = 27017,
+                 username: str = None, password: str = None):
         self.client = MongoClient(host=host, port=port, username=username, password=password)
 
     def read_db(self,
-                database_name='mydatabasename',
-                collection_name='mycollectionname',
-                query=(dict(), {'_id': 0})):
+                database_name: str = 'DBfastrader',
+                collection_name: str = 'financial',
+                query: Tuple[dict, dict] = (dict(), {'_id': 0})):
 
         db = self.client[database_name]
         collection = db[collection_name]
         start_time = time.perf_counter()
         df = pd.DataFrame(list(collection.find(*query)))
-        print('read data from mongo cost:{:.2f}s'.format(time.perf_counter() - start_time))
+        print('Read data from mongo cost:{:.2f}s'.format(time.perf_counter() - start_time))
         return df
 
-    def write_df_list_to_db(self,
-                            df,
-                            database_name='mydatabasename',
-                            collection_name='mycollectionname',
-                            delete_old_collection=True,
-                            chunksize=None,
-                            timer=True):
-        # """
-        # This function take a list and create a collection in MongoDB (you should
-        # provide the database name, collection, port to connect to the remoete database,
-        # server of the remote database, local port to tunnel to the other machine)
-        #
-        # ---------------------------------------------------------------------------
-        # Parameters / Input
-        #    my_list: the list to send to MongoDB
-        #    database_name:  database name
-        #    collection_name: collection name (to create)
-        #    chunksize: The number of items of the list that will be send at the
-        #        some time to the database. Default is None.
-        # ----------------------------------------------------------------------------
+    def write_df_dict_to_db(self,
+                            df: pd.DataFrame,
+                            database_name: str = 'mydatabasename',
+                            collection_name: str = 'mycollectionname',
+                            delete_old_collection: bool = True,
+                            chunksize: int = None,
+                            timer: bool = True):
+        """
+        将 dataframe 转成 list[dict] 后写入 mongodb, 如果df的col有重名, 在转dict过程中会丢失列
+        :param df:
+        :param database_name:
+        :param collection_name:
+        :param delete_old_collection:
+        :param chunksize:
+        :param timer:
+        :return:
+        """
 
         # To connect
         db = self.client[database_name]
@@ -91,40 +91,43 @@ class MongoDB(object):
                 database_name, collection_name))
         return
 
-    def write_df_json_to_db(self,
-                            my_df,
-                            database_name='mydatabasename',
-                            collection_name='mycollectionname',
-                            timer=True):
-        '''
-        三种写入数据的方式中最快, 快10多秒钟, 其他两种方式有chunksize的效果一般
-        缺点是 df.to_json(orient='records', date_format="iso") 以后,
-        再json.loads 生成dict里面的date会变成字符串, 需要加钩子函数去遍历列转成标准python的datetime格式, 影响效率
-        :param my_df: Dataframe
-        :param database_name: str
-        :param collection_name: str
-        :param server: str
-        :param mongodb_port: int/str
-        :param timer: Boolean
-        :return: None
-        '''
 
-        db = self.client[database_name]
-        collection = db[collection_name]
+def write_df_json_to_db(self,
+                        my_df: pd.DataFrame,
+                        database_name: str = 'mydatabasename',
+                        collection_name: str = 'mycollectionname',
+                        delete_old_collection: bool = True,
+                        timer: bool = True):
+    '''
+    三种写入数据的方式中最快, 快10多秒钟, 其他两种方式有chunksize的效果一般
+    缺点是 df.to_json(orient='records', date_format="iso") 以后,
+    再json.loads 生成dict里面的date会变成字符串, 需要加钩子函数去遍历列转成标准python的datetime格式, 影响效率
+    :param my_df: Dataframe
+    :param database_name: str
+    :param collection_name: str
+    :param delete_old_collection: Boolean
+    :param timer: Boolean
+    :return: None
+    '''
+
+    db = self.client[database_name]
+    collection = db[collection_name]
+
+    if delete_old_collection:
         collection.delete_many({})  # Destroy the collection
 
-        if timer:
-            start_time = time.perf_counter()
+    if timer:
+        start_time = time.perf_counter()
 
-        collection.insert_many(json.loads(my_df.to_json(orient='records')))
+    collection.insert_many(json.loads(my_df.to_json(orient='records')))
 
-        if timer:
-            print('Complete importing data to DB:{} -- collection:{}. Time cost : {:.2f}s'.format(
-                database_name, collection_name, time.perf_counter() - start_time))
-        else:
-            print('Complete importing data to DB:{} -- collection:{}.'.format(
-                database_name, collection_name))
-        return
+    if timer:
+        print('Complete importing data to DB:{} -- collection:{}. Time cost : {:.2f}s'.format(
+            database_name, collection_name, time.perf_counter() - start_time))
+    else:
+        print('Complete importing data to DB:{} -- collection:{}.'.format(
+            database_name, collection_name))
+    return
 
 
 if __name__ == '__main__':
@@ -142,10 +145,10 @@ if __name__ == '__main__':
                 + df['code'].apply(lambda x: '{:0>6d}'.format(x))
     print(f"Read csv data cost: {time.perf_counter() - start_time:.2f}s")
 
-    # pd_mongo.write_df_list_to_db(df, collection_name='from_df_list_nochunk')
-    # pd_mongo.write_df_list_to_db(df, collection_name='from_df_list_chunksize100', chunksize=5000)
+    # pd_mongo.write_df_dict_to_db(df, collection_name='from_df_list_nochunk')
+    # pd_mongo.write_df_dict_to_db(df, collection_name='from_df_list_chunksize100', chunksize=5000)
     write_start_time = time.perf_counter()
-    pd_mongo.write_df_list_to_db(df, database_name="DBfastrader", collection_name='financial', chunksize=1000)
+    pd_mongo.write_df_dict_to_db(df, database_name="DBfastrader", collection_name='financial', chunksize=1000)
     print(f"Write csv data cost: {time.perf_counter() - write_start_time:.2f}s")
 
     # dff = pd_mongo.read_db(database_name="DBfastrader", collection_name='financial')
